agents_backend/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "langchain-core>=0.3.34",
    "langchain-groq>=0.2.4",
    "langchain-mongodb>=0.4.0",
    "langgraph>=0.2.70",
    "langgraph-checkpoint-mongodb>=0.1.0",
    "opik>=1.4.11",
    "pre-commit>=4.1.0",
    "pydantic-settings>=2.7.1",
    "pymongo>=4.9.2",
    "loguru>=0.7.3",
    "langchain-huggingface>=0.1.2",
    "langchain-community>=0.3.17",
    "wikipedia>=1.4.0",
    "ipykernel>=6.29.5",
    "pydantic>=2.10.6",
    "datasketch>=1.6.5",
    "python-dotenv>=1.0.1",
    "livekit-agents[deepgram,elevenlabs,groq]~=1.4",
    "livekit-plugins-openai>=1.2.6",
    "livekit-plugins-deepgram>=1.2.6",
    "livekit-plugins-cartesia>=1.2.6",
    "livekit-plugins-silero>=1.2.6",
    "livekit-plugins-turn-detector>=1.2.6",
    "livekit-plugins-noise-cancellation>=0.2.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.7.2",
]

[tool.pip]
extra-index-url = "https://download.pytorch.org/whl/cpu/torch_stable.html"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]

[tool.ruff]
target-version = "py312"
